Update docs.en.md
I believe after testing that the "&" should actually be a "?".   There is a lot of chatter on the web about why this format does not work. People are finding ways around it like using the page url in place of the short page code. But if we correct the manual,(assuming I am right) we can save a lot of people the problems I have had all day.