Add explicit path to `sudo chown -R`
Given that it's possible a relatively novice user could end up on this page I feel it's important to have the path in the `chown` command. 

It's also best practice for something as potentially dangerous as `sudo chown -R` in case someone fat-fingers the command while scrolling their history and wipes out something important, like `/`.